Added functionality to spring_layout
Added an optional input argument, 'node_weight' to spring_layout. Given a string that is a node property this will treat it as the mass of the node for the purposes of calculating the anti-gravity force used in spring_layout. If none is given, it will continue to calculate it in the default manner. 

I find that when drawing certain graphs with large variations in the degrees of member nodes it is helpful for low-degree nodes to tend towards the edge of the drawing. This results in the high-degree nodes being pulled apart from each other as they are dragged along by their low-degree nodes, making the graph more readable.